requirements/base.txt
Django==2.2
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-js-asset==1.2.2
django-solo==1.1.5
install==1.3.4
Pillow==8.1.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
transliterate==1.10.2
